Moving student add/edit to it's own page for a nice layout.
Broke the StudentEditForm up. There's now a top level app with the
header elements bound to it. It contains the form code. This keeps
a better separation of concerns. I would also like to break the edit
form up a bit as it's getting quite messy.

Added new routes to the router file for all the new student endpoints.
There's now a student/new and student/edit/{id}

I've yet to do the edit page and update the student list page to
go to the new edit page instead of doing a pop up.