Add a modal popup warning user when leaving page with changes
Create a trackchanges module to track changes on a form. If the
user tries to leave the view it will propmpt them with a warning.
This does some ugly stuff with capturing all link clicks and also
injecting into the router navigate code. This isn't ideal but
should work for now. We should probably look at something even
more robust going forward.