Fix bug in bootstrap wysihtml5 plugin
There was a bug in the wysi plugin that would break uploading an
image without giving the textbox focus first. Now triggering the
focus on the textbox when the insert image or link button is
clicked. Also updated the template to autofocus the title field
when the page is rendered.